Specify platform specific font family & fallback font families
This change helps resolve emoji rendering issues which have been reported
by several different users across platforms. I filed two "canonical" issues
for this problem, but others have been reported.

- On Android: https://github.com/Fintasys/emoji_picker_flutter/issues/171
- On iOS: https://github.com/Fintasys/emoji_picker_flutter/issues/170

This change helps address this problem by setting a default font family on iOS,
MacOS, and Android. On iOS this is 'Apple Color Emoji', and on Android it is
'NotoColorEmoji'. We also specify several common fallback emoji font families
which Flutter will use if the specified font family is unavailable. This is
useful in case the user specifies an unavailable font family.

On iOS this actually resolves the rendering issue for U+2639, which you can see
in the description of #170. Unfortunately this change does not seem to resolve
the issue on Android when I tested it in the emulator, but it may help for some
configurations of Android where Flutter can access the 'NotoColorEmoji' font.

FIXED=#170
BUG=#171,#163